Move languages logics out of repository
